refactor(CodeEditor): update imports for CodeEditorModules type
- Changed imports of `CodeEditorModules` from `useModuleLoaders` to `moduleLoaders.types` across various hooks for improved clarity and consistency.
- Updated the export statement for `CodeEditorModules` in the hooks index file to align with the new import structure.
