+++ packages/code-editor/src/CodeEditor/hooks/extensions/useAutoCompleteExtension.ts
@@ -1,7 +1,7 @@
 import { type EditorView } from '@codemirror/view';
 
 import { type CodeEditorProps } from '../../CodeEditor.types';
-import { type CodeEditorModules } from '../useModuleLoaders';
+import { type CodeEditorModules } from '../moduleLoaders.types';
 
 import { useExtension } from './useExtension';
