+++ packages/code-editor/src/CodeEditor/hooks/extensions/useCodeFoldingExtension.tsx
@@ -7,7 +7,7 @@ import Icon from '@leafygreen-ui/icon';
 import { spacing } from '@leafygreen-ui/tokens';
 
 import { type CodeEditorProps } from '../../CodeEditor.types';
-import { type CodeEditorModules } from '../useModuleLoaders';
+import { type CodeEditorModules } from '../moduleLoaders.types';
 
 import { useExtension } from './useExtension';
