+++ packages/code-editor/src/CodeEditor/hooks/useModuleLoaders.ts
@@ -3,7 +3,7 @@ import { useMemo } from 'react';
 import { type CodeEditorProps } from '../CodeEditor.types';
 
 import { LanguageName } from './extensions/useLanguageExtension';
-import { CodeEditorModules } from './moduleLoaders.types';
+import { type CodeEditorModules } from './moduleLoaders.types';
 import { type LoadersMap } from './useLazyModules';
 
 /**
